remove some left apis in fluid.nn
