+++ python/paddle/static/nn/__init__.py
@@ -41,8 +41,6 @@ from .common import layer_norm  # noqa: F401
 from .common import embedding  # noqa: F401
 from ...fluid.contrib.layers import sparse_embedding  # noqa: F401
 from ...fluid.layers import StaticRNN  # noqa: F401
-from ...fluid.layers.nn import _pull_sparse  # noqa: F401
-from ...fluid.layers.nn import _pull_sparse_v2  # noqa: F401
 
 from .sequence_lod import sequence_conv  # noqa: F401
 from .sequence_lod import sequence_softmax  # noqa: F401
